backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
pydantic
pydantic-settings
pandas
pyyaml
python-dotenv
